+++ src/services/properties/deletePropertyById.js
@@ -1,12 +1,25 @@
 import {PrismaClient} from "@prisma/client";
 
+const prisma = new PrismaClient();
+
 const deletePropertyById = async (id) => {
-	const prisma = new PrismaClient();
-	const property = await prisma.property.delete({
-		where: {id},
-	});
+	try {
+		const property = await prisma.property.findFirst({
+			where: {id},
+		});
 
-	return property ? property.id : null;
+		if (!property) {
+			return null;
+		}
+
+		const deletedProperty = await prisma.property.delete({
+			where: {id},
+		});
+
+		return deletedProperty;
+	} catch (error) {
+		console.error("Error deleting Property:", error);
+		throw error;
+	}
 };
-
 export default deletePropertyById;
